requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
autograd=1.5=pypi_0
autograd-gamma=0.5.0=pypi_0
backcall=0.2.0=pypi_0
black=21.4b2=pypi_0
blas=1.0=openblas
blinker=1.6.2=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.1.0=pypi_0
comm=0.1.3=pypi_0
configargparse=1.5.3=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
dash=2.9.3=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
detectron2=0.6=pypi_0
docker-pycreds=0.4.0=pypi_0
docutils=0.17.1=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
fire=0.5.0=pypi_0
flask=2.3.2=pypi_0
fonttools=4.39.3=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.0=pypi_0
hydra-core=1.0.5=pypi_0
idna=3.4=pypi_0
imageio=2.28.1=pypi_0
importlib-metadata=3.10.1=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.23.1=pypi_0
ipython=8.13.2=pypi_0
ipywidgets=8.0.6=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.2.0=pypi_0
jupyter-core=5.3.0=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libopenblas=0.3.21=h043d6bf_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
loguru=0.7.0=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
minkowskiengine=0.5.4=pypi_0
mplcursors=0.5.2=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
natsort=8.3.1=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
networkx=3.1=pypi_0
ninja=1.10.2.3=pypi_0
nomkl=3.0=0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.0.6=pypi_0
open3d=0.17.0=pypi_0
openblas-devel=0.3.21=h06a4308_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=pypi_0
platformdirs=3.5.1=pypi_0
plotly=5.14.1=pypi_0
plyfile=0.9=pypi_0
pointnet2=0.0.0=pypi_0
portalocker=2.7.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.4=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.10.6=haa1d7c7_1
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
pytorch-lightning=1.7.2=pypi_0
pytz=2023.3=pypi_0
pyviz3d=0.2.32=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=25.0.2=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
regex=2023.5.5=pypi_0
reliability=0.8.9=pypi_0
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
sentry-sdk=1.22.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=66.0.0=pypi_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
torch=1.12.1+cu116=pypi_0
torch-scatter=2.1.1=pypi_0
torchaudio=0.12.1+cu116=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.13.1+cu116=pypi_0
tornado=6.3.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
trimesh=3.21.5=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=pypi_0
volumentations=0.1.8=pypi_0
wandb=0.15.2=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.3=pypi_0
wheel=0.38.4=pypi_0
widgetsnbextension=4.0.7=pypi_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.9.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
